Refactoring services layer
