+++ BusinessAssistant-opendata/src/test/java/com/businessassistantbcn/opendata/controller/OpendataControllerTest.java
@@ -1,11 +1,12 @@
 package com.businessassistantbcn.opendata.controller;
 
-import com.businessassistantbcn.opendata.service.*;
+import com.businessassistantbcn.opendata.service.config.DataConfigService;
+import com.businessassistantbcn.opendata.service.config.TestService;
+import com.businessassistantbcn.opendata.service.externaldata.*;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.reactive.WebFluxTest;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
